Change structure to avoid `print` in Django commands
