--- koocook_core/management/commands/_loadingr/__init__.py
@@ -1,13 +1,0 @@
-from django.core.management.base import BaseCommand, CommandError
-
-
-class Command(BaseCommand):
-    help = 'load recipe data'
-
-    def handle(self, *args, **options):
-        self.stdout.write('calling loadingr')
-        from .._add_path import add_datatrans
-        add_datatrans()
-        from .load_ingredient import main
-        main()
-
