+++ koocook_core/management/commands/_loadrecipe/__init__.py
@@ -5,7 +5,6 @@ class Command(BaseCommand):
     help = 'load recipe data'
 
     def handle(self, *args, **options):
-        self.stdout.write('I was called')
         self.stdout.write('calling read_data')
         from .._add_path import add_datatrans
         add_datatrans()
